mvp for profile display and update ready, with actions and reducers
