+++ src/components/pages/Home/RenderHomePage.js
@@ -3,12 +3,18 @@ import { connect } from 'react-redux';
 
 function RenderHomePage({ userData }) {
   return (
-    <div className="flux">
-      <h1 className="user-info">Hi {userData.name}</h1>
-      <h2>
+    <div className="flux" style={{ textAlign: 'center' }}>
+      <p className="user-info">
+        <h1>Hi {userData.name}</h1>
+      </p>
+      <p>
         This will be the landing page after Login, we should probably do a
-        conditional rendering depending on the type of user
-      </h2>
+        conditional rendering depending on the type of user.
+      </p>
+      <p>
+        Groomer could be taken to his Shop page or something like that. Clients
+        could be taken to the Map page or something.
+      </p>
     </div>
   );
 }
