+++ src/components/pages/ProfileEdit/UserEdit.js
@@ -1,16 +1,13 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Form, Input, Button } from 'antd';
-import produce from 'immer';
 
-const ClientEdit = ({ user }) => {
-  const [groomerData, setGroomerData] = useState({});
-
-  const handleChanges = e => {
-    setGroomerData(
-      produce(groomerData, draft => {
-        Object.assign(draft, e);
-      })
-    );
+const UserEdit = ({ user, setProfileEdit, profileEdit }) => {
+  //onFinish is a Antd Form's function that comes integrated that takes all the data on the form
+  //and passes it as key value pairs nested inside an object, so just using the event that's returned
+  //will be enough to sent the data to redux. no need to even keep track of state "manually".
+  const handleFinish = e => {
+    setProfileEdit(false);
+    profileEdit(e, 'editing');
   };
 
   return (
@@ -20,7 +17,7 @@ const ClientEdit = ({ user }) => {
         <Form
           layout="horizontal"
           className="forms"
-          onValuesChange={handleChanges}
+          onFinish={handleFinish}
           style={{
             width: 400,
             display: 'flex',
@@ -31,13 +28,12 @@ const ClientEdit = ({ user }) => {
           <Form.Item name="email" label="Email">
             <Input />
           </Form.Item>
-          <Form.Item name="phone" label="Phone Number">
+          <Form.Item name="preferred_username" label="Username">
             <Input />
           </Form.Item>
-          <Form.Item name="pic" label="Profile Pic">
+          <Form.Item name="family_name" label="Other Info">
             <Input />
           </Form.Item>
-          <Form.Item label="Field"></Form.Item>
           <Form.Item>
             <Button
               type="primary"
@@ -53,4 +49,4 @@ const ClientEdit = ({ user }) => {
   );
 };
 
-export default ClientEdit;
+export default UserEdit;
