+++ src/components/pages/ProfilePage/GroomerProfilePage.js
@@ -23,13 +23,9 @@ const GroomerProfilePage = ({ user }) => {
             <div className="user-info" style={{ fontSize: '1.3rem' }}>
               <span>Email: {user.email} </span>{' '}
               <span>Username: {user.preferred_username}</span>
-              <span>Ohter Info: {user.family_name}</span>
+              <span>Other Info: {user.family_name}</span>
             </div>
-            <button
-              style={{ border: '1px solid red' }}
-              className="edit"
-              onClick={handleChanges}
-            >
+            <button className="edit" onClick={handleChanges}>
               Edit profile
             </button>
           </div>
