--- src/components/pages/ProfilePage/mockdata.js
@@ -1,20 +1,0 @@
-const mockdata = [
-  {
-    id: 1,
-    name: 'Domingo Ocampo',
-    email: 'joe@groomer.cl',
-    phone: 3454456365,
-    pic: 'profilepic.img.com',
-    type: 1,
-  },
-  {
-    id: 2,
-    name: 'Camila Mena',
-    email: 'cmaila@gmail.com',
-    phone: 57567562,
-    pic: 'profilepic.img.url.',
-    type: 2,
-  },
-];
-
-export default mockdata;
