+++ src/styles/global.scss
@@ -143,4 +143,9 @@ svg {
   height: 40px;
   font-size: 1.3rem;
   font-weight: 600;
+  border: 1px solid black;
+  &:hover{
+    background: #001529;
+    color: white;
+  }
 }
